Extract Union functionality into extra class UnionType
